tidb: move out TxnState to seperate file from session.go
session.go file is large enough, move something out make it cleaner.
Anyway, TxnState is just another kv.Transaction, so the file is txn.go
